+++ test/Main.hs
@@ -12,6 +12,7 @@ import Pudu.Frontend.Lexer.ScannerSpec (scannerProperties)
 import Pudu.Frontend.ParserStateNameSpec (parserStateNameProperties)
 import Pudu.Frontend.ParserImportSpec (parserImportProperties)
 import Pudu.Frontend.ParserBindingSpec (parserBindingProperties)
+import Pudu.Frontend.ParserBlockSpec (parserBlockProperties)
 import Pudu.Frontend.ParserExpressionSpec (parserExpressionProperties)
 import Pudu.Frontend.ParserTypeSpec (parserTypeProperties)
 import Pudu.Frontend.SyntaxSpec (syntaxProperties)
@@ -49,9 +50,10 @@ main = do
   parserStateNameOutcomes <- traverse (uncurry check) parserStateNameProperties
   parserImportOutcomes <- traverse (uncurry check) parserImportProperties
   parserBindingOutcomes <- traverse (uncurry check) parserBindingProperties
+  parserBlockOutcomes <- traverse (uncurry check) parserBlockProperties
   parserTypeOutcomes <- traverse (uncurry check) parserTypeProperties
   parserExpressionOutcomes <- traverse (uncurry check) parserExpressionProperties
-  unless (and (sourceOutcomes <> diagnosticOutcomes <> tokenOutcomes <> cursorOutcomes <> scannerOutcomes <> numberSymbolOutcomes <> quotedOutcomes <> lexerOutcomes <> syntaxOutcomes <> parserStateNameOutcomes <> parserImportOutcomes <> parserBindingOutcomes <> parserTypeOutcomes <> parserExpressionOutcomes)) exitFailure
+  unless (and (sourceOutcomes <> diagnosticOutcomes <> tokenOutcomes <> cursorOutcomes <> scannerOutcomes <> numberSymbolOutcomes <> quotedOutcomes <> lexerOutcomes <> syntaxOutcomes <> parserStateNameOutcomes <> parserImportOutcomes <> parserBindingOutcomes <> parserBlockOutcomes <> parserTypeOutcomes <> parserExpressionOutcomes)) exitFailure
 check :: String -> IO Property -> IO Bool
 check label loadProperty = do
   putStrLn ("[test] " <> label)